mysite/requirements.txt
django-jet==1.0.7
rest_framework
